AC DC Sweep Work Fine
One hazard resolved in this step is the ac stamping for inductor, since
in the mainfram design, a branch is reserved for the inductor but in AC
analysis, actually it's not needed. So some trick need here. My trick is
do some transform to the MNA, detail please refer to the doc and the
readme.md