Single Diode Fine
Single diode can well converge.
However, in this step i change the API of the engine a little, that's,
state_definer() no longer need ANS as a input. That's because, in this
function, i somehow use np.insert to add a 0 at the very beginning of
the local_ans varible. This behavior accidentally somehow change the
np.ndarray( i believe it create a new object), so i had to explicitly
return this object so that the outsider can perceive. This eliminate the
usage of ANS pass( i actually put a lot effort to debug this, very
obscure)

However it a bit strange i Still need ANS passed into the engine,
because: 1) the engine has no sense of data-type, so it somehow need to
be decide outside 2), when using update rather than generator for
sweeping, use the original ANS can rid the trouble of decide whether
it's the first iteration.

I still want you know that the design in consistent, just improvised a
bit. take a note.