Fix a BUG of Doing Sweep and Converge Simultaneously
Previous version, in the state_definer(), clearing the last value of
converge_related item is combind with adding new value. However, when
the iteration exit, the last value will not be cleared and will be
carried into the new iteration of the next sweep value.

Current break the add and clear into seperature parts and all will be
done in a single loop, thus eliminate the problem above.

Test Ok when do a sweep on a RC circuit