Repair a Minor BUg in Inductor TRAN Analysis
for dc and op stamping of inductor, a small change should be add.
the stamping is equal to stamp a voltage meter or simple a wire.
Otherwise, a singular matrix error would be raised

However, after doing a double check on the algorithm, it's acutully
not compulsory to reserve a branch for the inductor. for the current
through can be represent as$$I(t1)=dt/L * U + I(t0)$$, where dt is the
time increment. But, using a new branch do is mroe convenient, becuase
thid method eliminate the need to calculate the current
everytime(because, the current is needed for the next time's stamping,
you do need to calculate it somehow)